Correction of some left-over this problems from previous check-in.
Addition of navier stokes equations in spherical coordiantes and
a number of test problems. 

Addition of a framework for introducing limiters to the DG framework
and also for permitting the use of implicit timesteppers, but why would you
want to?

Modification of the numerical flux used in the euler equations has
required re-generation of all the test data. This is still not properly
validated, but is in the demo's for testing purposes.

Addition of a weighting factor to include any geometric jacobian terms in
Z2 error calculations in non-carteisan coordiante systems and the 
appropriate regeneration of test data.

Corrected a bug so that the PeSt is now passed to the children of 
advection-diffusion and generalised advection-diffusion elements.

Modified the clamped_shell_with_arclength_cont.cc to include all three
pressure components and tweaked the parameter so that it converges. 
There is a strange problem on hector that Jonathan has discovered, but I
haven't looked into this yet. The output is funny, but the solution is 
correct.



git-svn-id: svn+ssh://maths1/home/svn-oomphlib/public/trunk@14 595e1d82-adf1-4282-bac7-b6073fa2ca8e