A number of modifications to merge and combine ElementWithExternalElements and FSIWallElements, unifying storage for equation numbering schemes and finite-difference routines for supplying entries to the Jacobian.
FiniteElements now inherit directly from GeomObjects and the position()
and, newly renamed, dposition_dt() functions call 
FiniteElement::interpolated_x() and FiniteElement::interpolated_dxdt() by
default. In addition, a new FiniteElement::zeta_nodal() function has been 
provided that interpolates the "global" intrinsic coordinate of the 
FiniteElement when viewed as a compound GeomObject in a Mesh of GeomObjects,
the default interpretation. The assumption is ALWAYS that zeta is 
interpolated using the nodal shape functions and at the FiniteElement level
zeta is chosen to be the Eulerian coordinate. In SolidFiniteElements
zeta is chosen to be the Lagrangian coordinate. In FaceElements, zeta is
chosen to be the boundary coordinate for the boundary of the mesh from which
the face is created. There is then an ambiguity for SolidFaceElement which
is resolved by choosing zeta to be the FaceElement representation in these
cases. The locate_zeta() function using Newton's method is now implemented
once in the FiniteElement base class using the zeta_nodal representation
or MacroElement representation is a MacroElement is defined.

ElementWithExternalElement have now been moved into their own files
and have been rewritten to use raw C-style storage underneath, rather than
vectors of vectors of vectors. Generic equation numbering routines, finite
difference routines for the Jacobian, etc have been unified in these 
elements. All external field data and external geometric data are stored in
separate containers, rather than as external data!

Hopefully, all documentation has been updated inline with these changes.

Minor changes to the routines that assign local equation numbers so that
the required overloading happens more automagically and can also be 
automagicked in ElementWithExternalElement.


INTERFACE CHANGE: The function GeomObject::drdt() is now 
called GeomObject::dpositiondt() for consistency.



git-svn-id: svn+ssh://maths1/home/svn-oomphlib/public/trunk@48 595e1d82-adf1-4282-bac7-b6073fa2ca8e