Unrefinement across a periodic boundary that lead to a periodic hanging node was broken in the case when the "original" node was deleted. The clear_copied_pointers() function in BoundaryNode has been modified so that rather than clearing the pointers, new memory is allocated and the values copied over; i.e. a shallow copy is turned into a deep copy before deleting the original.
In addition, the "pointer to a pointer" storage scheme for 
equation numbers in Data is not needed any more because periodic nodes are 
handled by a simpler approach. Periodicity in Solid Mechanics is enforced by
Lagrange multipliers. The Data storage scheme has been changed to a single 
pointer to equation numbers, saving a pointer per degree of freedom.


git-svn-id: svn+ssh://maths1/home/svn-oomphlib/public/trunk@104 595e1d82-adf1-4282-bac7-b6073fa2ca8e