Modified BlockPitchForkLinearSolver so that it works in distributed problems and with continuation. Note that the default parallel solver SuperLU_Dist does not handle the near-singular matrices as well as serial SuperLU. Thus, the Newton solver convergence degrades for the new parallel demo: track_pitch.cc. This problem can be remedied, by not using the Block decomposition or switching to the MUMPS solver.
The mumps solver interface has been modified to be
consistent with the latest LinearAlgebraDistribution changes.

Other small changes have been made to allow the 
BlockPitchForkLinearSolver to work in parallel by adding Halo_dofs
to the Problem (if needed) and modifying the get_derivative..
and get_hessian... functions to work on distributed problems.



git-svn-id: svn+ssh://maths1/home/svn-oomphlib/public/trunk@130 595e1d82-adf1-4282-bac7-b6073fa2ca8e