Check-in part-way through addition of 3D adaptive remeshing capability. A demo is included in navier_stokes that computes the flow past an icosahedron.
Interfaces to tetgen have been written and tetgen 1.4.3 is now included in 
external_src. There is an odd segmentation fault that occurs when tetgen is 
compiled with full optimisation. The same is true for triangle, so the 
Makefile.am has been hacked in tetgen. It has not yet in triangle because I 
want to get this in, but expect a small update soon to remedy this.

Wherever possible 2D (triangle) and 3D (tetgen) functionality has been 
harmonised and more efficient routines have been used to avoid excessive
use of maps when building meshes from scaffolds and building the scaffold
meshes themselves. Note that we have always required boundary information
for nodes, but not said so. Now I'm saying so. Each boundary node must be
supplied with an index that indicates one boundary on which it lies.

New integration schemes have been added for the TCrouzeixRaviartElements to
ensure full integration of the mass matrix. The enrichment of the 3D case to
quartic functions (2D is cubic) requires a much higher-order integration 
scheme which slows things down a lot, but that's the price of accuracy. 
If validation data has changed, it has been updated.

A few little fixes to include std::'s and other cross-complier issues have
also been fixed.



git-svn-id: svn+ssh://maths1/home/svn-oomphlib/public/trunk@188 595e1d82-adf1-4282-bac7-b6073fa2ca8e