Corrected an oversight when refinining structured meshes on curvilinear boundaries in 3D. Spotted by Phil Haines & David X. (Jim Denier's PhD Student). The problem was that Nodes that were hanging on boundaries did not get their position updated to be consistent with a macroelement representation when they became unhanging.
I have fixed the problem by keeping track of the nodes for which this could
occur in TreeBasedRefineableMesh::adapt_mesh() and then correcting them
afterwards. The same problem could arise with Lagrangian coordinates in
SolidMeshes and because this is hard to think about (and we don't have a 
test case), I've settled for simply throwing a warning.

A new test for this is included in 
demo_drivers/navier_stokes/three_d_entry_flow/full_tube.cc




git-svn-id: svn+ssh://maths1/home/svn-oomphlib/public/trunk@248 595e1d82-adf1-4282-bac7-b6073fa2ca8e