Deemed dump/restart for triangle meshes whose boundary lookup schemes have been manually messed with (e.g. in the constructor of a mesh that's derived from a TriangleMesh) to be "impossible" (for now). Removed corresponding test from
  demo_drivers/navier_stokes/unstructured_fluid/unstructured_two_d_fluid.cc

and added suitable paranoid test to restart function. to catch such cases
and issue a meaningful error message. The issue is that when dumping the
triangulateio data structure to disk, the manually altered boundary
lookup schemes will not have been incorporated into it. Given that this
is quite hacky anyway (and in the example mentioned above is only done
because the mesh was initially generated with xfig -- not really recommended
for serious applications) it not anticipated to be a major problem.
If it is, somebody will have to sort this out by updating the
triangulateio data structure before dumping...



git-svn-id: svn+ssh://maths1/home/svn-oomphlib/public/trunk@325 595e1d82-adf1-4282-bac7-b6073fa2ca8e