The analytical/chain-rule based evaluation of shape derivatives is quite broken -- certainly for refineable elements (with hanging nodes!). All the machinery needs to be carefully re-examined (see below).
For now we have kept the default method to be

  ElementWithMovingNodes::Method_for_shape_derivs =
    ElementWithMovingNodes::Shape_derivs_by_direct_fd

and have altered the functions

 void ElementWithMovingNodes::evaluate_shape_derivs_by_chain_rule(...)
 void ElementWithMovingNodes::evaluate_shape_derivs_by_fastest_method(...)

to throw an error, unless called with a new optional boolean argument,
called i_know_what_i_am_doing, that can be used by "expert users".
This currently done in a few demo driver codes where we deliberately
test this functionality and it works "OK", either because the elements
aren't refineable, or the degregation of the Newton method isn't
bad (in fact, it generally isn't anyway).

The function

 void ElementWithMovingNodes::evaluate_shape_derivs_by_direct_fd()

was kept intact since it just re-asserts the default behaviour 
and is therefore harmless.

The problem (there may be more!) seems to be the lack of distinction
between the element's nodes and its shape controlling nodes -- they
can differ for refineable elements with hanging nodes.

What we (Matthias and Patrick) know is that

   ElementWithMovingNodes::get_dnodal_coordinates_dgeom_dofs(...)

does the right thing (though, interestingly, the comment at the
top of the function does not explicitly spell out that
the j in dX_{ij} / d s_l refers to the element's shape-controlling rather than
its own nodes. Whoever takes this up must work their way backwards
through the machinery starting from

  ElementWithMovingNodes::fill_in_jacobian_from_geometric_data(...)

which also seems OK. Specifically, the rank-three tensor
dresidual_dnodal_coordinates[] in that function is dimensioned (and then
used) correctly when the chain rule is being evaluated, BUT all
the elements that actually implement (or overload)
get_dresidual_dnodal_coordinates(...) seem to do this wrongly.
The call/inheritance sequence for this function is quite deep but
it seems that few/none of the implementations use (or can use) the
required "shape controlling node" machinery provided by
RefineableElement::nshape_controlling_nodes() and
RefineableElement::shape_controlling_node_lookup() since much of the
machinery is implemented at the level of non-refineable elements 
(which obviously don't have access to these functions).

 Note also that some of the specific implementations of
get_dresidual_dnodal_coordinates(...) are wrong for other reasons.
For instance, RefineablePoissonEquations::get_dresidual_dnodal_coordinates(...)
refers to psi(q) (in the shape derivative of the source fct term)
where q loops over the shape controlling nodes -- there can be
more of those than the number of nodes in the element in which case
we'd get a nasty range checking error (or seg fault).

More horrors may lurk elsewhere....




git-svn-id: svn+ssh://maths1/home/svn-oomphlib/public/trunk@397 595e1d82-adf1-4282-bac7-b6073fa2ca8e