Modified: matrices .cc and .h
matrices.cc: fixed a bug with a PARANOID test which can be made to
always fail. I incorrectly used || (logical OR), it should be be &&
(logical AND).
Added a lot more PARANOID tests for concatenate_without_communication(...)
It now can accept a DenseMatrix of pointers to CRDoubleMatrix which may
be null. So we do not have to create a zero matrix, it is now 100%
backwards compatible with build_preconditioner_matrix(...)

matrices.h: added more to the comments.

git-svn-id: svn+ssh://maths1/home/svn-oomphlib/public/trunk@438 595e1d82-adf1-4282-bac7-b6073fa2ca8e