src/generic/block_preconditioner.h
* Moved the creation of the distributions in
Precomputed_block_distribution_pt[] to block_setup(...) - this is where the
distributions in Block_distribution_pt[] is also created.
If any distibutions in Precomputed_block_distribution_pt already exist,
we delete them.

* In the destructor, we delete the distributionsi in
Precomputed_block_distribution_pt[], then clear the Vector. I tried to
delete it in clear_block_preconditioner_base(...), but it just won't
work... not sure why. Doing it this way is equivalent, combined with the
modification above, this is probably more robust.

git-svn-id: svn+ssh://maths1/home/svn-oomphlib/public/trunk@459 595e1d82-adf1-4282-bac7-b6073fa2ca8e