This is a check-in to ensure that the new complex_harmonic eigenproblem is included. The demo includes complex eigenvalues and eigenvectors and is even documented as well (although this may need some revision/improvement).
Also coming in are modifications to the boundary conditions in
the multi_domain_boussinesq_convection demo so that it will work with 
either TaylorHood or CrouzeixRaviart Elements without other changes.

The changes to src/generic and the brethreton.cc demo results from the 
inclusion of an option to create and store vectors of pointers to dofs within
each element. This is not normally required, but is needed in the revised 
formulation of continuation/bifurcation tracking problems. 
This is still "work in progress". 

However, it has lead to an INTERFACE CHANGE:

 All functions that assign LOCAL equation numbers now take a boolean flag 
that determines whether the pointers to LOCAL dofs should be stored within 
the element. If this affects you at all it will be because you need to change
assign_local_eqn_numbers() to include the boolean option in your custom 
elements. This change was required in the bretherton.cc demo_driver, so look 
into that one if you are unsure.





git-svn-id: svn+ssh://maths1/home/svn-oomphlib/public/trunk@711 595e1d82-adf1-4282-bac7-b6073fa2ca8e