Removing legacy code from block_preconditioner.cc:
modified:   src/generic/block_preconditioner.cc

Attempting to remove all legacy code to do with
not being able to re-classify dof types
(done at the element level).

Commented out the block of code which checks that
if row has been imported the block number is the same.

A bit of history: A long time ago, the block preconditioner
will loop through the meshes and get all the global dof numbers,
along with the associated dof type numbers via the function
GeneralisedElement::get_dof_numbers_for_unknowns(...) and put the
results in a set of pairs. This is the comment from a old
commit of mine:

==========================================================================
$ git log -S"hen a FaceElement will re-classify the bulk element's DOFs"
.
commit 3935ae6f98fc45e8313edbbc75ba1be50c40795d
Author: mly <mly@595e1d82-adf1-4282-bac7-b6073fa2ca8e>
Date:   Mon Dec 2 14:55:49 2013 +0000

    Fixed distributed mesh problem in generic/block_preconditioner.cc

    The old code used a set of pairs to store the global equation and
dof
    type pair. If a dof is re-classified, then this is also stored.
    This the set is used later to get the number of dofs and such.

    This means we have more dofs then there actually is. So, we used a
map
    instead.
==========================================================================

This is the code commented out block_preconditioner.cc:1514

         {
          unsigned u = first_dof_to_send[p] + ndof_to_recv[p];
          for (unsigned i = first_dof_to_send[p]; i < u; i++)
           {
 #ifdef PARANOID
            // RAYRAY Legacy code, this checks that the DOFs have not
            // been
            // re-classified. This has been changed so that we can
            // re-classify
            // the DOFs. Example: Bulk element classifies it's own DOF
            // types,
            // then a FaceElement will re-classify the bulk element's
            // DOFs.
            // See ImposeParallelOutflowElement.
            // I have not deleted the below so we can easily "revert"
            // back if
            // required.

            // paranoid check
            // check that if row has been imported the block number is
            // the
            // same
//            if (dof_recv[my_global_dofs[i]-first_row])
//             {
//              if (Dof_number_dense[my_global_dofs[i]-first_row]
//                  != my_dof_numbers[i])
//               {
//                std::ostringstream error_message;
//                error_message
//                 << "Inconsistency in assigment of block numbers\n"
//                 << "Global dof " <<  my_global_dofs[i]
//                 << "was previously assigned to block "
//                 <<  Dof_number_dense[my_global_dofs[i]-first_row]
//                 << "\nNow it's been reassigned to block "
//                 << my_dof_numbers[i] << ".\n"
//                 << "This is most likely because one of your\n"
//                 << "elements has classified its degrees of freedom\n"
//                 << "wrongly. You should remember that \n"
//                 <<
//                 "GeneralisedElement::get_block_numbers_for_unknowns()\n"
//                 << "should only classify the element's OWN dofs and
//                 not \n"
//                 << "deal with dofs that were created by resizing
//                 nodes, \n"
//                 << "say. Check that loops over nodal values in that
//                 \n"
//                 << "function do not use Node::nvalue() to determine
//                 the\n"
//                 << "dofs to be classified!\n";
//                throw OomphLibWarning(error_message.str(),
//                                      OOMPH_CURRENT_FUNCTION,
//                                      OOMPH_EXCEPTION_LOCATION);
//               }
//             }
            // indicate that this dof has ben recv
            dof_recv[my_global_dofs[i]-first_row] = true;
 #endif
            Dof_number_dense[my_global_dofs[i]-first_row] =
             my_dof_numbers[i];
           }
         }

As you can see, it loops through the dofs, if it has been received,
it checks that the dof type number is the same as the one defined in
the my_dof_numbers array. The arrays my_dof_numbers and my_global_dofs
form a pair of vectors which contains the global dof types and dof types
numbers accordingly.

SO, in the old code (before the commit above, using a set instead of a
map),
if we re-classify global dofs, (say, global dof 5 from 3 to 6) we have
the
following vectors:

index:            0,1, 2 ,3, 4 ,5,6
my_global_dofs = [3,4,*5*,7,*5*,8,9]
my_dof_numbers = [1,2,*3*,1,*6*,3,2]

The above commented out loop will catch this since my_dof_numbers[4] is
6,
but  Dof_number_dense[5] is already assigned 3 (via a previous loop
iteration).

HOWEVER! - when using a map and just changing the value (dof type
number), the
vectors are now:

index:            0,1,2, 3 ,4,5
my_global_dofs = [3,4,7,*5*,8,9]
my_dof_numbers = [1,2,1,*6*,3,2]

As you can see, only the last visited dof classification is kept.
So the check for re-classification of dof types will always pass,
since there is no trace of re-classification.

Thus this paranoia check is safe to take out.

HOWEVER! - I have left the array dof_recv in since it's used to check
if every local dof has been received.

git-svn-id: svn+ssh://maths1/home/svn-oomphlib/public/trunk@1066 595e1d82-adf1-4282-bac7-b6073fa2ca8e