Added 3D surfactant demo, which solves the Rayleigh--Plateau problem in three dimensions in the presence of an insoluble surfactant. This is a "hard" test because the system is formulated in Cartesians, but remains axisymmetric.
In addition, the (old) insoluble_surfactant and soluble_surfactant demos
have been updated to include the curvature terms properly in the new
formulation. These are two-dimensional (non-axisymmetric) problems, so the
complete range of equations for insoluble surfactant are now tested. 

Documentation will be written for the next submission, including the 
general theory ... watch this space.



git-svn-id: svn+ssh://maths1/home/svn-oomphlib/public/trunk@1083 595e1d82-adf1-4282-bac7-b6073fa2ca8e