Yet more changes to the FluidInterface formulation
A generic SurfactantTransportInterfaceElement class has
been added to test the generic functionality and (of course)
certain interfaces have had to be changed accordingly. 

More functions have been merged into the base FluidInterfaceElement
class. The surface derivatives are now implemented in separate
classes that are introduced into the inheritence hierarchy as a  
template parameter in the final specialised functions. 

The 3d_rayleight_instability_surfactant.cc demo has now
been rewritten to use these generic elements, but the 
axisymmetric version still implements in place, as before.

The Elastic version of the surfactant elements has not yet
been tested. This will feature in the next check-in I hope.



git-svn-id: svn+ssh://maths1/home/svn-oomphlib/public/trunk@1088 595e1d82-adf1-4282-bac7-b6073fa2ca8e