Development for book chapter about adaptive bifucation tracking.
Added continuous formulation of the linearised Navier--Stokes equations to 
the libraray *NOT FULLY TESTED* and wrote a demo driver that computes
the base flow and eigenproblem on separate meshes. The demo appears to work,
but considerably more testing needs to be done in order to understand the
best type of error measure to use on each mesh.

This required one additional function interpolated_dudx to be added to the
vanilla Navier--Stokes elements.

NOTE: I have had to delete the demo_drivers/unsteady_heat/*_midpoint 
directory because there was a tree clash. I don't know why, but if this
needs to be reinstated then please feel free to use an earlier revision
to get hold of the files.



git-svn-id: svn+ssh://maths1/home/svn-oomphlib/public/trunk@1313 595e1d82-adf1-4282-bac7-b6073fa2ca8e