Small changes to make the code work with gcc8. The problems were all inheritence-related  and the solutions were that specific versions of overloaded functions needed to be explicitly called, or final overloads provided.
Note that there is also a problem if difference versions
of mpfr happens to be installed on your machine and you
don't ask to use it. The problem is essentially due to how
shared libraraies work in Linux and there is no easy
fix that I have found. The solution is to make sure that
the LD_LIBRARY_PATH in your local shell has an 
explicit path to oomph-lib's installation location of 
mpfr.



git-svn-id: svn+ssh://maths1/home/svn-oomphlib/public/trunk@1316 595e1d82-adf1-4282-bac7-b6073fa2ca8e